Move the IContactRegisterService to the core
